Update - fetch next cdc reference using db query builder - generate next cdc reference for regular and testing campaigns - Remove unused code for sortByMultiple - Delete unused migrations - Add timestamps to has_fields table - Create data migration

+++ database/seeders/MediaTableSeeder.php
@@ -18,17 +18,26 @@ class MediaTableSeeder extends Seeder
     public function run()
     {
         $media = DB::table('media')->select('attachable_id', 'attachable_type', 'id')->get();
+        $totalMedia = $media->count();
+        $totalNewMediaInserted = 0;
+
         foreach ($media as $file) {
-            DB::table('has_media')->insert([
+            $inseted = DB::table('has_media')->insert([
                 'attachable_id' => $file->attachable_id,
                 'attachable_type' => $file->attachable_type,
                 'media_id' => $file->id
             ]);
+
+            if ($inseted) {
+                $totalNewMediaInserted += 1;
+            }
         }
 
-        Schema::table('media', function (Blueprint $table) {
-            $table->dropColumn('attachable_id');
-            $table->dropColumn('attachable_type');
-        });
+        if ($totalMedia == $totalNewMediaInserted) {
+            Schema::table('media', function (Blueprint $table) {
+                $table->dropColumn('attachable_id');
+                $table->dropColumn('attachable_type');
+            });
+        }
     }
 }
